1. Moved download manager logic to a separate file
